requirements.txt
numpy
Pillow
torch==2.3.0
transformers==4.41.2
gradio==4.37.1